src/consumer/requirements.txt
google-cloud-datastore
tweepy
google-cloud
google-cloud-pubsub